ml_ops/113/jobs/requirements.txt
black~=22.0
flake8>=3.7.9, <5.0
isort~=5.0
pydantic~=2.7
abeja-sdk~=2.2
python-json-logger~=2.0
streamlit~=1.35
PyYAML~=6.0
requests~=2.26
python-dotenv~=1.0
Pillow==9.2.0
scipy==1.9.2
networkx
tqdm
tensorboardX
imageio~=2.22.0
